Move engine selection out of main
